refactor(cmd): move templates package under internal/
Relocate pkg/cmd/templates to pkg/cmd/internal/templates. The package
has zero consumers outside pkg/cmd/, so Go's internal/ rule formalizes
the boundary that already exists in practice and prevents future
accidental external imports of CLI rendering helpers.

Mechanical change: every consumer in pkg/cmd/commands/ updates the
import path from "pkg/cmd/templates" to "pkg/cmd/internal/templates".
The package name stays "templates", so call sites are unchanged.
