+++ pkg/cmd/commands/progress.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/block/schemabot/pkg/cmd/client"
-	"github.com/block/schemabot/pkg/cmd/templates"
+	"github.com/block/schemabot/pkg/cmd/internal/templates"
 )
 
 // ProgressCmd watches schema change progress for a specific apply.
