+++ pkg/cmd/commands/settings.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	"github.com/block/schemabot/pkg/cmd/client"
-	"github.com/block/schemabot/pkg/cmd/templates"
+	"github.com/block/schemabot/pkg/cmd/internal/templates"
 )
 
 // settingInfo holds metadata for a known setting.
