+++ pkg/cmd/internal/templates/locks.go
