+++ pkg/cmd/internal/templates/plan.go
