+++ pkg/cmd/internal/templates/preview.go
