+++ pkg/cmd/internal/templates/preview_defer.go
