+++ pkg/cmd/internal/templates/preview_progress.go
