+++ pkg/cmd/internal/templates/preview_sequential.go
