+++ pkg/cmd/internal/templates/preview_vschema.go
