+++ pkg/cmd/internal/templates/progress.go
