+++ pkg/cmd/internal/templates/progress_parse.go
