+++ pkg/cmd/internal/templates/progress_render.go
