+++ pkg/cmd/internal/templates/progress_shard.go
