+++ pkg/cmd/internal/templates/progress_shard_test.go
