+++ pkg/cmd/internal/templates/progress_states_test.go
