+++ pkg/cmd/internal/templates/time.go
